requirements.txt
discord.py==2.3.1
Flask==2.2.2
python-dotenv==1.0.0
urllib3==1.26.12